[TASK] Adapt name "TYPO3 CMS"
The official name of TYPO3 is "TYPO3 CMS" now. Appropriate changes made.

A note in chapter "About this document" explains the focus on TYPO3 CMS
and clarifies when the document means TYPO3 as the project and not the
CMS. Whenever the guide refers to a TYPO3 CMS specific functionality,
"TYPO3 CMS" is used.

Resolves: #42646

Change-Id: I7cadef4a724bb4e0404a073e139f71d4a7aed763