[TASK] Update screenshots to show new TYPO3 logo in backend
The TYPO3 logo and its colours have been changed. The TYPO3 Style Guide
is available at: http://typo3.org/the-brand/style-guide/

This change had an impact on the TYPO3 logo used in the TYPO3 CMS backend
and therefore new screenshots have to be made. The image file names were
changed slightly during the process as well.

Resolves: #42631

Change-Id: Ie277b6678b55cf8a47afcd73c65d77a306503be2