[TASK] Eliminate "generated" headlines
In the OpenOffice document we sometimes left out one level of headlines.
During the migration to ReST, however every level of headlines, if it
existed or not, was written in the .rst files; also the missing ones.
These headlines now show up as "((generated))" in the files and in the
rendered output.

Replace each "generated" headline with the one which should actually be
there.

Change-Id: I71f9be873f15c37bb092d517cc5cc023960e4e8c
Resolves: #45892
Reviewed-on: https://review.typo3.org/18523
Reviewed-by: Christopher
Tested-by: Christopher