[TASK] Add advice to remove old code after a security update
The Security Guide should recommend that administrators should remove
any old and possibly insecure code from the server after a security
update. This applies to the TYPO3 core as well as TYPO3 extensions.
This patch introduces a new section in chapter "General Guidelines",
which describes the risk for TYPO3 core updates, and expands the
section "Keep TYPO3 extensions up-to-date".

Resolves: #37722
Change-Id: I5c7ec5c434f20a6ae3a5d4ecdd7d23e5aebafa04
Reviewed-on: https://review.typo3.org/24542
Reviewed-by: Michael Schams
Tested-by: Michael Schams