[TASK] Remove superfluous and misleading info about width of tabs
The CGL contained information about the assumed width of tabs, which
made some users think, indenting with spaces(!) would be OK as long as
the correct number of spaces were used. This is not the case. Also was
the information on the width of tabs superfluous. It therefore is
removed.

Change-Id: Ia4b00845c71fc59291ae1698f6006517f01cee3a
Resolves: #53976
Reviewed-on: https://review.typo3.org/28566
Reviewed-by: Chris topher
Tested-by: Chris topher