[TASK] Salted passwords for BE user and Install Tool
This patch updates the documentation and mentions the use of salted MD5
hashes as the Install Tool password since TYPO3 CMS 6.2 and where to find
it (file LocalConfiguration.php). This applies to chapter "Guidelines for
TYPO3 Integrators -> Install Tool".

It also replaces the related screenshot, which shows the new Install Tool
of TYPO3 CMS 6.2.x now.

Resolves: #53026
Change-Id: If208b37e07f26b690c9935e0e1bc6e8cd80cb939
Reviewed-on: https://review.typo3.org/28629
Reviewed-by: Michael Schams
Tested-by: Michael Schams