[TASK] Add latexpdf settings to Settings.yml
With these settings a normal Sphinx-build for LaTeX-PDF is possible.
For details see the manual of the "sphinx" extension, chapter "Rendering
on docs.typo3.org", section "PDF rendering".

Also fix some broken directives and stop using a nested table, which
would result in this error:
Markup is unsupported in LaTeX:
JavaScript/UsingExtjs/Viewport/Index:: nested tables are not yet
implemented.