[DOC] Fix English issues in extension scanner documentation
Phew... a lot of "germanisms" in this one. I've corrected a lot of the sentence structure and terminology to sound more English, as well as several typos. I also homogenised the use of ReST vs. RST and removed quotation marks where they were used around RST. I've trimmed several sections which stated the same thing twice or included information that did not pertain to the subject of the section.

Tips for the author:

* Try not to use this many quotation marks. Green, strong, weak *really* do not lend themselves to quotation; doing that indicates you gave them new meaning which you didn't, so it only serves to confuse the eyes of the reader.
* Be careful about spaghetti sentences and syntax. For example, the proper English syntax would be "to easily eat a whole pie" rather than "to eat a whole pie easily" which sounds wrong to English speakers.
* There were some redundant sentence bits. Obviously, the core team decides if a matcher should be categorised differently and there's no need to describe the entire decision process. I've trimmed down that part of the documentation significantly.
* English uses far fewer commas.