Highlight set() behaviour for easier understanding
Adds an emphasis to the part of the text explaining the set() method and the auto-creation of a named parameter as it is quite hard to spot after most of the documentation tells you to use 'createNamedParameter' whenever possible. Adds an additional info box to the createNamedParameter paragraph explaining the set() behaviour.