TASK: Keep old implementation of update wizards
Even if V9 provides a modern API, the old is still valid. As we also
keep documentation for ext_update, we should also keep this
documentation.

A hint was added to use the newer API, still if someone has to work with
old implementations, the documentation is still valid and helpful.