[FEATURE] Clarify backend module api
- use formatted lists for parameters to make it more readable
- add additional information about parameters
- add information about how to configure modules with TypoScript
- add note about backend actions not being cached

There is currently a nearly identical page about "Backend Modules"
in the extbase book. Since this page and the one in the extbase book
overlap in most parts, the page in the extbase book can be
taken down and just link to "TYPO3 Explained". For this reason, we
add the part about TypoScript and the note about backend actions
not being cached, so the information will not be lost.