Add chapter "TypoScript Syntax"
For versions 9.5 and above most parts of the
separate manual "TypoScript syntax and in depth study"
were merged into TYPO3.

The older parts (<= 8.7) remained in this separate manual.
This resulted in confusion as it was not possible to switch
between versions from 8.7 and below to 9.5 and above (and
vice versa) for individual pages.

As moving the (older) information to "TYPO3 Explained" would
have been considerable work to sort (because not everything
is moved to T3 Explained, some parts to TSref etc.), we
copied the information from 9.5 to older versions.

Some of this should be checked if consistent for older
versions (e.g. conditions), but most of the information should
be the same across versions.