Restructure configuration section
Pull everything up to the top menu level in the configuration
chapter. That makes it easier to navigate and we can unify the
toctree in the main Index.rst (previously we displayed the
contents with a level of 2 for all and a level of 3 for configuration.
Now we can change all to level of 2 which makes maintainance easier).