[TASK] Fix inconsistend handling of StringFrontend
In version 9, StringFrontend was merely deprecated while it was
removed entirely from 9.5 docs.

In 10, it was removed, but was not removed from the docs.

Instead, when deprecated, a note about deprecation should be added
and a link to the changelog, while it should get removed when removed
from core.

Related: #889
Releases: master, 10.4, 9.5