Remove duplicate information about deprecation
How do deprecate classes, methods etc. in the TYPO3 core is handled in
the contribution guide. The section in the CGL chapter covers basically
the same thing. This text is removed and a link to the contribution guide
added.

The chapter "deprecation" handles the topic deprecation from the point
of view of people who use TYPO3. It covers how to disable deprecation
notices, how the deprecation log works and how to find calls to deprecated
functions in extensions.

Resolves: #910