Restructure: Links and references
Links and references are 2 different things but there is some overlap.
This commit groups the following subchapters together:

- Linkhandler API
- LinkBrowser API
- Softreferences

into a chapter "Links & references".

As next step we could add an introduction page explaining concepts
of references in TYPO3, how links are created and stored internally,
and also possibly how the reference index works.

Related is also the FAL which is covered in its own chapter, but
a link to that could be added in the general section.

The advantage is that there are less top-level menu items and more
related topics grouped together.