Added note regarding command definitions
I noticed that in the TYPO3 docs the commands are explicitly defined, while the Symfony docs say commands are automatically added due to `autoconfigure: true`. To avoid confusion for Symfony newbies like myself, I propose to add a little note about why `autoconfigure: true` is not enough in this case.