Change overview for upgrade wizard
On the implementation page an overview over necessary steps is displayed.
This way, the reader can see all the necessary steps in one place before
looking at the details.

Also, the necessary step of registering is moved up before the optional
interfaces.

Because an overview over the interfaces is now on the implementation page,
it is dropped from the start page in the chapter (Index.rst).