Add naming conventions for upgrade wizard identifier
Note: This convention does not depend on a specific core version. It
can be used as of now, in core and third party extensions for TYPO3 11
and older versions alike. But other identifiers which do not adhere to
this convention will still work and continue to work in the future.

Resolves: #1334
Releases: master, 10.4