Clarify extension-key section in composer.json
Some text is obsolete, because specifying the extension
key is now mandatory.

Additionally, the example was not clear, because you might easily
interpret is as meaning the extension key with underscore replaced
by dash instead of the actual string "extension-key".

We now use the same example as in
https://github.com/TYPO3/CmsComposerInstallers

Releases: master, 10.4, 9.5