Make it more clear that FluidEmail and MailMessage (#1381)
Since everything is on one page, it is often easy to overlook if there
are 2 different methods for doing one thing.

Also, readers may be directed directly to a subsection (via the anchors)
and then miss the context.

This change adds some extra explanation clearly pointing out that
FluidEmail and MailMessage are 2 different methods to do one thing.