[FEATURE] Add link to crowdin on publish page (#1424)
* [FEATURE] Add link to crowdin on publish page

In order to publish an extension, another step is (usually)
necessary: setup the translation on Crowdin. This is already
described in the "Localization" chapter. A small teaser text
and a link is now added to the "Publish an Extension" page.

* Commit suggested change by review

Co-authored-by: Mathias Brodala <mbrodala@pagemachine.de>

* Commit suggested change by review

Co-authored-by: Mathias Brodala <mbrodala@pagemachine.de>

Co-authored-by: Mathias Brodala <mbrodala@pagemachine.de>

(cherry picked from commit 4cf1b2085543857d97a11b83fca01477dd8a13b7)