[TASK] Restructure information about links
A chapter "Links & references" is started to combine all information about links,
references and relations.

This now contains the following (already existing) pages:
- linkhandler API
- linkbrowser
- soft references.

Later, more information can be added, including an explanation of differences
between references, soft references, links and relations and including a definition
of terms like "linkhandler", "typolink" and a description of the "t3 URI syntax".

To avoid broken links, either redirects can be added or "orphan" pages replacing
the original location of the pages and a link to the new pages. To enable switching
back and forth from older versions, the "orphan" pages can be preferred and can be
removed in later versions.

Related: #1619
Releases: master, 11.5