Use plaintext instead of bullet list
In this case, the use of a bullet list does not provide any
advantage. Also, the longer information following the last
API bullet point ("Use these API functions") was awkward
as the function decription was a part of the bullet point
but the formatting indicated otherweise.

As this is not a very important change, backport is not
necessary, but can be done.

Releases: main