Remove wrong information about security severities (#1631)
The information about the security severities (critical, etc.) is removed.
It was incorrect as the severity indicates the impact of an exploitation of
a vulnerability, not how likely the breach can be.
This was formerly misleading.

Resolves: #1224
Releases: master, 11.5, 10.4, 9.5
(cherry picked from commit 07a31908983c70a3340fca7951a1a9ed804c5fc2)