Remove the "One level higher" from the paths
When using an own extension to acquire the JS files, we do not need to go one one level higher to get the file. If we add the two dots (../) then the path looks like this:

<script type="text/javascript" charset="utf-8" async="" data-requirecontext="_" data-requiremodule="leaflet" src="./../typo3conf/ext/my_ext/Resources/Public/JavaScript/leaflet.js?bust=1641223806"></script> 

and this redirects to the main Backend page and as a result it response with an HTML header instead of a JavaScript program. 

If we remove the two dots and keep the tracing slash, TYPO3 responds with the right header and the JavaScript works again