Expand string description (#1813)
Elaborate on when to use the 3 string types in YAML, and why.  This also normalizes the rules to be more in line with common practice both in TYPO3 and elsewhere.  (I checked and could find no examples of YAML conventions that said to quote everything; YAML bare strings are almost universally accepted.)

cf:
https://review.typo3.org/c/Packages/TYPO3.CMS/+/74326
https://forge.typo3.org/issues/86896