Prefer usage of LoggerInterface
In the logging section, the LoggerInterface (available since v11)
is used in some sections, but also LoggerAwareTrait and
LoggerAwareInterface are used.

This change prefers the usage of instantiation the logger via
dependency injection and removes usage of LoggerAwareTrait.

Additionally, a note is added to mention LoggerAwareTrait and
refer to older versions of this page.
Since the LoggerAwareTrait is still heavily used in the core of
v12 and not yet deprecated, it seems prudent to add this note
as orientation.