Remove .. only:: html directives
These directives used to be important since we had HTML and PDF
rendering but the PDF rendering has been dropped.

So we no longer use these directives in general.

Also on one of the pages, the indenting was not correct: All
content to which the directive applies should be indented and
it was not.

Related: TYPO3-Documentation/T3DocTeam#11