Use term "custom content element type"
The page "Creating a custom content element" explains how to create a "custom
content element type", not how to create a "content element". So to better
reflect this, the title is changed and "type" added.

In the text, we already use the phrase "content element type" in some places,
this is changed in some other places as well.

This makes the title and text more clear and correct.

Similar changes where already made in the TypoScript Reference where we have
objects and object types and it was sometimes used interchangeably (and
incorrectly).

Related: #TYPO3-Documentation/T3DocTeam#77