Clarify backend module signature
Add and explain the term "backend module signature".

Add examples for backend module signatures.

Fix error in explanation of backend module signature (not extkey
and module key is used, but main module key and submodule key).

When referring to the configuration section in the backend, add
the name of the module function "Backend modules" where the listing
of modules can be found.

Resolves: TYPO3-Documentation/TYPO3CMS-Reference-TSconfig#126