Add naming conventions for mm tables
The naming conventions page already contains several rules
and examples for database tables.

A section is now added for mm tables which contain records
for multiple-multiple relations between tables.

The rules for Extbase are clear.

For non-Extbase no such rules currently exist, but it is
suggested to recommend a similar rule as in the Extbase
tables.