Add note about language records (#1893)
In this [StackOverflow question](https://stackoverflow.com/questions/72467813/typo3-10-4-how-to-add-available-language), a 10.4 user had difficulties to add new languages to the site configuration because he was not aware that sys_language records are necessary to do so.

I've added a hint to to the 10.4 documentation. Maybe that that hint should also be applied to these pages:
- https://docs.typo3.org/m/typo3/reference-coreapi/10.4/en-us/ApiOverview/SiteHandling/Basics.html#sitehandling-basics
- https://docs.typo3.org/m/typo3/tutorial-editors/10.4/en-us/Languages/Index.html#languages
- https://docs.typo3.org/m/typo3/guide-frontendlocalization/10.4/en-us/SettingUpLanguages/Index.html

Feel free to improve my wording!