[TASK] Fix explanation of fourth parameter of FlashMessage
Part of the explanation was duplicate and the second explanation
was under the severities and not within the explanation of the
constructor for FlashMessage.

Additionall, the names of the arguments are used instead of
line numbers, which makes it more clear and prevents problems
in case the snippet is modified or arguments removed.

Releases: main, 11.5