Move page "Create custom ViewHelper" from Extbase book (#1953) (#1954)
* [TASK] Move Develop Custom ViewHelper from Extbase book

The page "Developing a custom ViewHelper" was copied from
the Extbase book.

The text will be optimized in subsequent pull requests

Related: https://github.com/TYPO3-Documentation/TYPO3CMS-Book-ExtbaseFluid#536

* Reduce content for custom ViewHelper

The previous text for developing a custom ViewHelper used
several steps to add more and more functionality to the
ViewHelper.

While such an approach often makes sense, here this was
applied very fine granular, with many steps.
This resulted in a lot of code duplication.

So, the multiple step approach was kept, but the first few
steps were merged into one code snippet with the various
parts explained in the following sections.
Having the complete code snippet in one place, makes it
shorter and easier to read (and has less code duplication).

* Further optimization of custom ViewHelper page