Replace term object "object accessors"
This term was used in the Extbase book to refer to
variables, but it is not really used in the Fluid
world. In the typo3fluid documentation, "expressions"
are used.

Also, the term object accessors is somewhat misleading
as in the examples, the variable might not refer to
an object.