[TASK] Remove additional Fluid links
In the Introduction page, there was an additional section with
links to external sources. This is now mostly redundant, as
more information has been included in the Fluid chapter. For
example the page "Syntax" was included, so it is no longer
necessary to link to external source. Also, some of the links
are duplicates with the links already included on the page
"Using Fluid in TYPO3".

Related: TYPO3-Documentation/TYPO3CMS-Book-ExtbaseFluid#536