[TASK] Add layout example for Fluid
Fluid introduction was a bit thin, as Fluid layout was explained,
but an example was missing. We have now added a very minimal
example.

Since this part, should replace the page "Creating a consistent
look and feel with layouts" in the Extbase book, it should include
an example.

Related: TYPO3-Documentation/TYPO3CMS-Book-ExtbaseFluid#536