Point out possible problems with redis cache backend
Point out problems which may occur in case not enough memory
is allocated for the Redis server. If an eviction policy is
used, the eviction of entries may result in inconsistent
data due to cache tag related entries missing, which could
result in problems of cache for a page not being flushed at
all by "normal" measures (such as when content is changed or
a flush of the page cache is requested).

Resolves: #2050