Create a tutorial section in the link browsers
Differnciate into a tutorial that works and you can follow and the explanaition of API.

Remove the news example and link to the news documentation where it is described in detail instead.

Use the Haiku record example from the examples extension instead.

On the fly: blog example switched from master to main so the composer.json had to be adjusted.

releases: main, 11.5