Create a tutorial section in the link browsers (#2284)
* Create a tutorial section in the link browsers

Differnciate into a tutorial that works and you can follow and the explanaition of API.

Remove the news example and link to the news documentation where it is described in detail instead.

Use the Haiku record example from the examples extension instead.

On the fly: blog example switched from master to main so the composer.json had to be adjusted.

releases: main, 11.5

* Create a tutorial section in the link browsers

Differnciate into a tutorial that works and you can follow and the explanaition of API.

Remove the news example and link to the news documentation where it is described in detail instead.

Use the Haiku record example from the examples extension instead.

On the fly: blog example switched from master to main so the composer.json had to be adjusted.

releases: main, 11.5

* Generate Code snippets

* Apply suggestions from code review

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>

* Update Documentation/ApiOverview/LinkBrowser/Tutorials/RecordLinkBrowser.rst

* Add Screenshot

Co-authored-by: lina.wolf <lwolf@w-commerce.de>
Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>

(cherry picked from commit d0a9ed15b673de124a49f9a529a48ae0cc76a804)