Add clarification for how to use log channels
The documentation for the channels contains mostly the
information from the changelog which does not explain
how the channel can be used in the logging
configuration.

While on closer look, this may become obvious, there
have been questions raised.

Also, in the logging configuration, the further examples
are put into one code block, another example is added
with the class name and a formatting is used
to hightlight the composition of the array keys in
$GLOBALS['TYPO3_CONF_VARS']['LOG']. Formatting it
this way should make it much clearer, how the array
keys are composed, what is always the same and what
contains the component name:

$GLOBALS['TYPO3_CONF_VARS']['LOG']
  ['Documentation']['Examples']['Controller']['FalExampleController']
  ['writerConfiguration'] = [
      // ...
  ];

Resolves: #2315