[FEATURE] Apply code-quality checks for PHP
Apply CLI and linting checks to the PHP files contained in this manual.

We are starting to use php files now and include them see ExtensionArchitecture/HowTo/Localization/_php for example. These now get checked for CLI and linted so we don't have to do it manually. As a bonus we also check our internal configuration files and apply checks to the composer.json

Code quality checks are integrated in our CI