[FEATURE] Apply code-quality checks for PHP (#2322)
* [FEATURE] Apply code-quality checks for PHP

Apply CLI and linting checks to the PHP files contained in this manual.

We are starting to use php files now and include them see ExtensionArchitecture/HowTo/Localization/_php for example. These now get checked for CLI and linted so we don't have to do it manually. As a bonus we also check our internal configuration files and apply checks to the composer.json

Code quality checks are integrated in our CI

* Remove composer normalize etc and rector

* [TASK] Try to resolve CGI-testing errors on github actions

* FIX CLI run

* Tests fixen

* [TASK] Add auth.json to gitignore

* [TASK] Apply cgl fixes to code files

* [TASK] Use dry-run for cgl check in GitHub Action workflow

* Apply suggestions from code review

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>

* Apply changes from review

* Use latest ubuntu version

Co-authored-by: lina.wolf <lwolf@w-commerce.de>
Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>