[TASK] Revise "Limit restrictions to tables" section
The examples for using restrictions only for a specific table are hidden at the bottom of the chapter. They are now moved to the according section which makes it easier for users to grab the examples.

Additionally, use a list for better readability of the restrictions in a specific QueryRestrictionContainer. Also add LimitToTablesRestrictionContainer to list of QueryRestrictionContainer.

Releases: main, 11.5