[BACKPORT][TASK] Revise "Limit restrictions to tables" section (#2481)
The examples for using restrictions only for a specific table are hidden at the bottom of the chapter. They are now moved to the according section which makes it easier for users to grab the examples.

Additionally, use a list for better readability of the restrictions in a specific QueryRestrictionContainer. Also add LimitToTablesRestrictionContainer to list of QueryRestrictionContainer.

Releases: main, 11.5

Co-authored-by: Lina Wolf <48202465+linawolf@users.noreply.github.com>

(cherry picked from commit 27ff8a8a7bf4825e060130e0b66ac31e631d6e64)