[TASK] Do not recommend to inject query builder via DI
Injecting the query builder is error-prone as it contains state.
The user must clone the object or reset the query parts to avoid
side effects. Therefore we should not recommend this and show as
best practise.

The query builder should be retrieved via ConnectionPool or
Connection object to avoid these errors.

The example uses the Connection class instead.

Releases: main, 11.5