[DEPRECATION] #99120 - TypoScriptParser
Add notes and todos about the deprecation where the old TypoScript parser is explained.
This PR is only about the deprecation, once the new TypoScript parser is stable these
chapters should be revised.

The transition from the ":=" hook to the PSR-14 event will be addressed in a separate
PR.

Related: https://github.com/TYPO3-Documentation/Changelog-To-Doc/issues/287
Releases: main