[Backport 11.5] [TASK] Move file list events code snippets into separate files (#2679)
This is the first PR in a series to move the code snippets into separate files. This way,
we can adapt coding styles easily and migrate PHP code to newer PHP versions via Rector
automatically.

- YAML configuration is moved into separate files
- PHP classes are moved into separate files
- Namespaces are adjusted (from EventListener -> FileList)
- Extension key is moved from "my_package" to "my_extension" to be more generic
  (and to avoid similarity with site package)

When this PR is merged, the other events are adjusted in the same manner.

Releases: main, 11.5