Add section $GLOBALS['LANG'] and amend information
A new section is added to the page on GLOBALS for
$GLOBALS['LANG'].

This section provides a very short introduction and
then links to the already existing page "Localization in PHP".

In the existing page "Localization in PHP" an introductory
text is added, which summarizes what the array does and
provides overview when it is inialized.

Resolves: #2737
Resolves: #2462