[TASK] List configuration files for site and system-wide
List the possible configuration files for a site (`config/sites/`) and system-wide (`config/system/`). This gives an integrator a better overview at this place which files are processed and what is their purpose.

This is a preparation for the new Content Security Policy which introduces a
`config/sites/<my_site>/csp.yaml` file.

Releases: main