Update links on composer.json page
When the page was originally created, several links were added to
resources by Helmut Hummel (who had done considerable work in this
area). However, the resources have become outdated, and in particular
configuration options which are now listed as deprecated in official
documenation are still used.

Thus, we removed the links.

Additionally, a link to the typo3/cms-composer-installers package
is added, as this is relevant for Composer.