Update Extension testing to chapter for ExtensionTestEnvironment
Previously, a snippet was added to composer.json using ExtensionTestEnvironment.
This is no longer recommended since TYPO3 v11 as the functionality
was moved to the core.

It is however necessary to have Public/Resources directory.

See https://review.typo3.org/c/Packages/TYPO3.CMS/+/71029

The functionality was responsible for making sure that the extension
was installed into .Build/Web/typo3conf/ext when setting up a test
environment.

Related: #2921