Further updates to Extension Testing
The PHP, TYPO3, testing-framework etc. versions are updated for
TYPO3 v11.

Where possible, versions are replace with some generic text to
reduce number of places where versions must be updated.

The general description of the branches of enetcache is kept.
Though that is outdated, it just explains the general principles
and does not have to be updated every time.

Related: #2921