Clarify if one or more core versions should be tested
Some additional changes are made to resolve inconsistency
in "Extension Testing" page regarding if one or more core
versions should be tested.

For now, we test only one core version (as done in original
version of this page). However, we add a note that it should
be decided in the future if one or more core versions should
be tested in this page.

We also point out the change in enetcache which switched from
testing one core version per branch to 2 versions per branch.

There are still some changes which need to be made, this
is pointed out in an "attention" note at the top of the page.