Use composerUpdate instead of composerInstall
In styleguide as well as enetcache, composerUpdate is used
instead of composerInstall.

This makes sense as it will install the latest versions, even
if a composer.lock already exists (e.g. when running the tests
locally).

Additionally, the list of features for runTests.sh is formatted
as list as this makes the individual points better readable
for skimming over the part.