[TASK] Enhance service configuration examples in PSR-14 event
To avoid mistakes when not familiar with dependency injection and
the configuration of a Services.yaml file, the Services.yaml examples
now have a comment to hint the user that the example file is only
an excerpt of a Services.yaml file. Furthermore, a link to the
according DI section is added for more information.

Releases: main, 12.4