Update testing extensions (11.5 branch) (#2926)
* Update Extension testing to chapter for ExtensionTestEnvironment

Previously, a snippet was added to composer.json using ExtensionTestEnvironment.
This is no longer recommended since TYPO3 v11 as the functionality
was moved to the core.

It is however necessary to have Public/Resources directory.

See https://review.typo3.org/c/Packages/TYPO3.CMS/+/71029

The functionality was responsible for making sure that the extension
was installed into .Build/Web/typo3conf/ext when setting up a test
environment.

Related: #2921

* Improve text style

Co-authored-by: Lina Wolf <48202465+linawolf@users.noreply.github.com>

* Remove replace / self.version in Extension Testing

Configuration in composer.json using replace, such as:

"replace": {
    "enetcache": "self.version",
    "typo3-ter/enetcache": "self.version"
  },

is no longer recommended in newer TYPO3 versions and may result in
warnings by Composer.

Related: #2921

* Remove cms-package-dir

Related: #2921

* Remove ignore-as-root

Since TYPO3 v11, it is no longer necessary to use
ignore-as-root in composer.json (and it has no effect).

Related: #2921

* Remove note

* Further updates to Extension Testing

The PHP, TYPO3, testing-framework etc. versions are updated for
TYPO3 v11.

Where possible, versions are replace with some generic text to
reduce number of places where versions must be updated.

The general description of the branches of enetcache is kept.
Though that is outdated, it just explains the general principles
and does not have to be updated every time.

Related: #2921

* Remove extra cms-package-dir

* Add formatting

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>

* Improve spelling

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>

* Add formatting

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>

* Add trailing slash for directory for more clarity

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>

* Clarify if one or more core versions should be tested

Some additional changes are made to resolve inconsistency
in "Extension Testing" page regarding if one or more core
versions should be tested.

For now, we test only one core version (as done in original
version of this page). However, we add a note that it should
be decided in the future if one or more core versions should
be tested in this page.

We also point out the change in enetcache which switched from
testing one core version per branch to 2 versions per branch.

There are still some changes which need to be made, this
is pointed out in an "attention" note at the top of the page.

* Remove outdated information for extension testing

Remove outdated information about testing with multiple TYPO3
versions

Related: #2972

* Additional update for Extension Testing

Remove section about installing to .Build/Web/typo3conf/ext
since that is no longer supported. Instead, the paths
should be updated so this is no longer necessary.

* Update core version in Extension Testing

* Use composerUpdate instead of composerInstall

In styleguide as well as enetcache, composerUpdate is used
instead of composerInstall.

This makes sense as it will install the latest versions, even
if a composer.lock already exists (e.g. when running the tests
locally).

Additionally, the list of features for runTests.sh is formatted
as list as this makes the individual points better readable
for skimming over the part.

* Use generic phrase for several PHP versions

Since this page will be updated in the future and the PHP
versions will change every time and sometimes 1, 2 or 3
are supported we use a generic phrase.

* Update test output

* Remove note on top

---------

Co-authored-by: Lina Wolf <48202465+linawolf@users.noreply.github.com>
Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>