[TASK] Remove RequireJS content
With #3177 the migration section for RequireJS to ES6 modules was removed. However, some reference are still there. They are now linked to the older docs version.

Additionally, as RequireJS is not available anymore in TYPO3 v13, the according chapters are removed from the navigation. The pages itself are left with only a note about the removal of RequireJS and a link to the migration in the 12.4 docs.

Related: https://github.com/TYPO3-Documentation/Changelog-To-Doc/issues/559
Releases: main