[Backport 12.4] [TASK] Add webhooks chapter (#3357)
In "A-Z" a new chapter "Webhooks and reactions" is added to guide the user to the with v12 introduced features. This makes them a little bit more prominent.

As there is currently no dedicated manual for the Webhooks system extension available, the changelog
is linked for the time being.

Releases: main, 12.4