[TASK] Move FAL introduction to main page (#3638)
The FAL main page is just a TOC. You have to click on "Introduction" to get more information about that. As that introduction page is small, the content is moved to the main page. This makes it easier for the user to grasp the content of that chapter.

Additionally, the relevant keywords are referenced to the appropriate chapter.

Releases: main, 12.4, 11.5