Explain term extkeyprefix (#3736)
On the top of the page "Naming conventions" some terms are
explained. Especially extension keys are complicated and
there are several variants in place, depending on where
it is used (e.g. as UpperCamelCase, with stripped away
underscores etc.).

The term often already reflects how the term is written,
e.g. ExtensionName in UpperCamelCase where the extension
key is UpperCamelCased'd.

We now explain the extkeyprefix (formerly extension-prefix)
used in the tables section.