[TASK] #102621 - Most TSFE members marked internal or read only
This change adds appropriate versionchanged directive to the changed properties of TSFE.

The AfterCacheableContentIsGeneratedEvent provides an example which uses the property "content" that is marked as internal now. A todo was added for now as a solution is not known by now.

The methods marked as internal will be taken care of in a follow-up.

Additionally, TSFE->baseUrl was removed in a previous Core patch. This property was already deprecated with v12. This change also addresses this change.

Related: https://github.com/TYPO3-Documentation/Changelog-To-Doc/issues/752
Releases: main