[TASK] Link to PageAccessFailureReasons to failure reasons
The `TSFE->getPageAccessFailureReasons()` method has been marked as internal with v13.0.
Independently of this, the reference to this method does not really help the user. Instead,
the `PageAccessFailureReasons` class holds the reasons for failures as constants.

This can safely be backported to the LTS versions as the class exists there.

Related: https://github.com/TYPO3-Documentation/Changelog-To-Doc/issues/752
Releases: main, 12.4, 11.5