[CI] Remove check for "important" admonition
Once upon a time, when the "important" admonition was coloured green. This was bad as it looks like "everything is okay". But an "important" admonition as - well - important.

But the times have changed: Nowadays the admonitions only know two colours: orange and dark grey. Therefore, this check is now superfluous and can vanish.

Releases: main, 12.4