[Backport 12.4][TASK] Remove speeddemo
It is not really best practice to use the sitepackage
builder to create an extension.

Also the content is outdated.

I would also suggest to discontinue and archive
https://github.com/TYPO3-Documentation/speeddemo

Releases: main, 12.4

* [TASK] Remove speeddemo

It is not really best practice to use the sitepackage
builder to create an extension.

Also the content is outdated.

I would also suggest to discontinue and archive
https://github.com/TYPO3-Documentation/speeddemo

Releases: main, 12.4

* [TASK] Display extension structure in a directory tree

(cherry picked from commit 8fc728f789149fdf0c96bb03143ccb0484763e3b)