[TASK] Unify and improve Services.yaml examples
As suggested by @schliesser in #4390 the example formerly
used in Documentation/Configuration/Yaml/ServicesYaml.rst was not close to real Life. Most extensions will need to exclude Domain/Model/* and an example how to exclude directories is generally helpful.

Further more I made the examples unified, theoretical and not bound to an extension and lintable by moving them to a yaml file.