[TASK Set a more suitable group identifier
Contrary to the specification of the tab identifier "common" in the configuration of the New Content Element Wizard, the group must be named "default" in order to actually be listed among the default elements in the CType Select field. Otherwise, a new select item group with the name "common" is created in the element list.