[DOCS] Add example to use 403 ErrorRedirectHandler for login redirects
This patch shows an example for the frontend login
how to get a redirect when an unauthorized visitor
tries to access a restricted page.

In this case, a 403 ErrorHandler can perform a
redirect with special parameters, that a properly
configured felogin plugin can utilize to
redirect back to.

Resolves: #104694
Releases: main, 12.4, 11.5
Change-Id: Ia2c6723621304b8b3de366590c1be905572b00d1