[DOCS] Document database records (#4760)
* [DOCS] Document database records

A detailed explanation of what they are was missing.

This is a preparation for documenting the
v13 additions of record objects.

Releases: main, 11.5

* Update Documentation/ApiOverview/Database/Index.rst

Co-authored-by: Stefan Frömken <froemken@gmail.com>

* Apply suggestions from code review

Co-authored-by: Stefan Frömken <froemken@gmail.com>

* Apply suggestions from code review

Co-authored-by: Stefan Frömken <froemken@gmail.com>

* Update Documentation/ApiOverview/DatabaseRecords/Index.rst

Co-authored-by: Stefan Frömken <froemken@gmail.com>

* Update Documentation/ApiOverview/Database/Index.rst

Co-authored-by: Stefan Frömken <froemken@gmail.com>

* Update Documentation/ApiOverview/Database/Index.rst

* Update Index.rst

* Update Index.rst

* [DOCS] Document database records

A detailed explanation of what they are was missing.

This is a preparation for documenting the
v13 additions of record objects.

Releases: main, 11.5

---------

Co-authored-by: Stefan Frömken <froemken@gmail.com>

(cherry picked from commit 480828c309c8cf59f7aa4f12a8b2422ae77224e4)